use type null(null) for ndnson nulls instead of string(null)
This commit changes the inferparse in the ndjson reader to
use type null for null values instead of type string.

When introspecting zng data using typeof(.), I found it odd
that fields with nulls were coding as type "string".  This is
especially confusing when the non-null fields are numbers.
It makes more sense to see type null and type float64 mixed
together for a numeric column instead of null and string.