refine semantic analaysis and fix bug with constants in filters
This commit further develops the semantic analysis component of the
compiler.  This was needed to fix a problem where the lift-filter
optimization did not carry with it any Z constants that were defined
in the Z program causing such constant refs to be treated as field
references in the filters generated by the zbuf.Filter methods.

The fix is to fix the intermediate represenation of the Z program
to explicitly track the global scope of the Z program and initialize
constants in the runtime zctx before any filters are created.  Also,
when distributed workers are generated and use the AsProc() method
to zbuf.Filter, the AST returned here needs to include said constants
so this has been fixed.

In doing this refactor, we took one step toward the work needed
to split the parser AST and semantic DSL into logically distinct
components as outlined in issue #2163.